move to internal tooltip and add disabled state
