+++ packages/gestalt/src/Tooltip.js
@@ -1,50 +1,7 @@
 // @flow strict
-import { type Node, useReducer, useRef } from 'react';
-import Box from './Box.js';
-import Controller from './Controller.js';
-import Layer from './Layer.js';
-import Text from './Text.js';
-import useDebouncedCallback from './useDebouncedCallback.js';
+import { type Node } from 'react';
+import InternalTooltip from './Tooltip/InternalTooltip.js';
 import { type Indexable } from './zIndex.js';
-
-const noop = () => {};
-const TIMEOUT = 100;
-
-const initialState = { hoveredIcon: false, hoveredText: false, isOpen: false };
-
-const reducer = (
-  state: {| hoveredIcon: boolean, hoveredText: boolean, isOpen: boolean |},
-  action: {| type: 'hoverInIcon' | 'hoverInText' | 'hoverOutIcon' | 'hoverOutText' |},
-) => {
-  switch (action.type) {
-    case 'hoverInIcon':
-      return {
-        ...state,
-        hoveredIcon: true,
-        isOpen: true,
-      };
-    case 'hoverInText':
-      return {
-        ...state,
-        hoveredText: true,
-        isOpen: true,
-      };
-    case 'hoverOutIcon':
-      return {
-        ...state,
-        hoveredIcon: false,
-        isOpen: !state.hoveredText ? false : state.isOpen,
-      };
-    case 'hoverOutText':
-      return {
-        ...state,
-        hoveredText: false,
-        isOpen: !state.hoveredIcon ? false : state.isOpen,
-      };
-    default:
-      throw new Error();
-  }
-};
 
 type Props = {|
   /**
@@ -94,74 +51,17 @@ export default function Tooltip({
   text,
   zIndex,
 }: Props): Node {
-  const [state, dispatch] = useReducer(reducer, initialState);
-  const { isOpen } = state;
-
-  const childRef = useRef<?HTMLElement>(null);
-  const { current: anchor } = childRef;
-
-  const mouseLeaveDelay = link ? TIMEOUT : 0;
-
-  const handleIconMouseEnter = () => {
-    dispatch({ type: 'hoverInIcon' });
-  };
-
-  const handleIconMouseLeave = useDebouncedCallback(() => {
-    dispatch({ type: 'hoverOutIcon' });
-  }, mouseLeaveDelay);
-
-  const handleTextMouseEnter = () => {
-    dispatch({ type: 'hoverInText' });
-  };
-
-  const handleTextMouseLeave = useDebouncedCallback(() => {
-    dispatch({ type: 'hoverOutText' });
-  }, mouseLeaveDelay);
-
   return (
-    <Box display={inline ? 'inlineBlock' : 'block'}>
-      <Box
-        aria-label={accessibilityLabel != null ? accessibilityLabel : text}
-        ref={childRef}
-        onFocus={handleIconMouseEnter}
-        onBlur={handleIconMouseLeave}
-        onMouseEnter={handleIconMouseEnter}
-        onMouseLeave={handleIconMouseLeave}
-      >
-        {children}
-      </Box>
-      {isOpen && !!anchor && (
-        <Layer zIndex={zIndex}>
-          <Controller
-            anchor={anchor}
-            caret={false}
-            bgColor="darkGray"
-            border={false}
-            idealDirection={idealDirection}
-            onDismiss={noop}
-            positionRelativeToAnchor={false}
-            rounding={2}
-            size={null}
-          >
-            <Box
-              maxWidth={180}
-              padding={2}
-              onBlur={link ? handleTextMouseLeave : undefined}
-              onFocus={link ? handleTextMouseEnter : undefined}
-              onMouseEnter={link ? handleTextMouseEnter : undefined}
-              onMouseLeave={link ? handleTextMouseLeave : undefined}
-              role="tooltip"
-              tabIndex={0}
-            >
-              <Text color="inverse" size="100">
-                {text}
-              </Text>
-              {Boolean(link) && <Box marginTop={1}>{link}</Box>}
-            </Box>
-          </Controller>
-        </Layer>
-      )}
-    </Box>
+    <InternalTooltip
+      accessibilityLabel={accessibilityLabel}
+      link={link}
+      idealDirection={idealDirection}
+      inline={inline}
+      text={text}
+      zIndex={zIndex}
+    >
+      {children}
+    </InternalTooltip>
   );
 }
